feat: consolidate MonadThrow and derivations into Monad
* derivations.ts: removed
* monad_throw.ts: removed
* monad.ts: inherits createMonad and MonadThrow
* types.ts: updated to reference new MonadThrow location
* testing: moved derivations.test.ts to monad.test.ts

Fixes #80
Fixes #101
